+++ src/main.py
@@ -26,8 +26,8 @@ def print_class(cls):
 
     outp += START + ("=" * 20) + "  " + cls["title"] + "  " + ("=" * 20) + END + "\n"
 
-    format_str = "{:^8} | {:^14} | {:^30} | {:^10} | {:^16} | {:^8} | {:^12} | {:^6} | {:^8}\n"
-    outp += format_str.format("Section", "Class Number", "Instructor", "Open Seats", "Reserved Seats", "Waiting", "Status", "Days", "Timespan")
+    format_str = "{:^8} | {:^4} | {:^14} | {:^30} | {:^10} | {:^16} | {:^10} | {:^8} | {:^12} | {:^6} | {:^8}\n"
+    outp += format_str.format("Section", "Type", "Class Number", "Instructor", "Open Seats", "Reserved Seats", "Taken", "Waiting", "Status", "Days", "Timespan")
     outp += "-"*144 + "\n"
 
     sections = cls["sections"]
@@ -43,9 +43,10 @@ def print_class(cls):
 
         if inst_cond and clos_cond:
             found_professor = True
-            outp += format_str.format(section['section'],
+            outp += format_str.format(section['section'], section["type"],
                 section['class_number'], section['instructor'], 
-                section['open_seats'], section['reserved_seats'], section['waiting'], 
+                section['open_seats'], section['reserved_seats'], 
+                section['taken'], section['waiting'], 
                 section['status'], section['days'], section['timespan'])
 
     if found_professor: