+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-# 11-08-2018
+# 11-09-2017
+- Added type and taken to the collected data
+- Made sorting sections work
+
+# 11-08-2017
 - Implemented error checking
 - Ensure that DOM is (a little more) ready before assuming table is there
+- Parsed all info into a Python dictionary
+- Pretty print info into a table
 
 # 11-07-2017
 - Implemented basic searching through PASS department selector/class table